Move :truck #5 : style 파일에서 .scss 파일들 이동
