+++ src/components/header/Header.scss
@@ -1,4 +1,4 @@
-@import '../variables';
+@import '../../styles/reset', '../../styles/variables';
 
 * {
   text-decoration: none;
@@ -54,6 +54,7 @@
       line-height: 20px;
       letter-spacing: -0.005em;
       text-align: center;
+      color: #000000;
     }
   }
 }
