+++ src/components/headerPost/HeaderPost.scss
@@ -1,4 +1,4 @@
-@import '../reset', '../variables';
+@import '../../styles/reset', '../../styles/variables';
 
 .hidden {
   display: none;
