+++ src/components/headerPost/components/CountPerson.scss
@@ -1,4 +1,4 @@
-@import '../reset', '../variables';
+@import '../../../styles/reset', '../../../styles/variables';
 
 .headerPost__profile-container {
   display: flex;
