+++ src/components/headerPost/components/DropDownEmojis.scss
@@ -1,4 +1,4 @@
-@import '../reset', '../variables';
+@import '../../../styles/reset', '../../../styles/variables';
 
 .headerPost__addition-emojis-container,
 .headerPost__dropdown-emojis-container {
