+++ src/components/headerPost/components/PostToast.scss
@@ -1,4 +1,4 @@
-@import '../reset', '../variables';
+@import '../../../styles/reset', '../../../styles/variables';
 
 .sharetoast {
   position: fixed;
